Upgrade GoCardless API and use gocardless_pro
GoCardless updated their API and deprecated the older one. These changes
keep the same donation flow with the new API.

Note that there's a slightly ugly relationship between the
GoCardlessHelper and the session. This is because we need to hand the
user off to GC to create a "customer" object (that contains the bank
details), and that object is accessible after being redirected.

It would be nice to find a neater way to do this, as right now
the logic is split between the helper class, the middleware and the
views.