Light updates
Coming to this fresh after a year or so, I find the 'about, not a venue for' line confusing. Not sure what it means, nor why it's in title case. (Sorry Sym, I know you liked it, but I'm not sure it makes sense... Democracy Club is about making democracy better...) 

The first line of the code of conduct should be the line explaining what it is. Then the header of 'applies' to everyone, then the text that explains everyone/everywhere. 

Then the rest is fine.